Add support for Alchemy 7

+++ Rakefile
@@ -1,28 +1,28 @@
 begin
-  require 'bundler/setup'
+  require "bundler/setup"
 rescue LoadError
-  puts 'You must `gem install bundler` and `bundle install` to run rake tasks'
+  puts "You must `gem install bundler` and `bundle install` to run rake tasks"
 end
 
-require 'rdoc/task'
+require "rdoc/task"
 
 RDoc::Task.new(:rdoc) do |rdoc|
-  rdoc.rdoc_dir = 'rdoc'
-  rdoc.title    = 'AlchemyDevise'
-  rdoc.options << '--line-numbers'
-  rdoc.rdoc_files.include('README.md')
-  rdoc.rdoc_files.include('lib/**/*.rb')
-  rdoc.rdoc_files.include('app/**/*.rb')
+  rdoc.rdoc_dir = "rdoc"
+  rdoc.title = "AlchemyDevise"
+  rdoc.options << "--line-numbers"
+  rdoc.rdoc_files.include("README.md")
+  rdoc.rdoc_files.include("lib/**/*.rb")
+  rdoc.rdoc_files.include("app/**/*.rb")
 end
 
 APP_RAKEFILE = File.expand_path("../spec/dummy/Rakefile", __FILE__)
-load 'rails/tasks/engine.rake'
+load "rails/tasks/engine.rake"
 
-require 'rspec/core/rake_task'
+require "rspec/core/rake_task"
 
 RSpec::Core::RakeTask.new(:spec)
 
-task :default => ['alchemy:spec:prepare', :spec]
+task :default => ["alchemy:spec:prepare", :spec]
 
 Bundler::GemHelper.install_tasks
 
@@ -30,26 +30,36 @@ namespace :alchemy do
   namespace :spec do
     desc "Prepares database for testing Alchemy"
     task :prepare do
-      system <<-BASH
-cd spec/dummy
-export RAILS_ENV=test
-bin/rake railties:install:migrations
-bin/rake db:drop db:create db:migrate
-bin/rails g alchemy:install --force --auto-accept
-bin/rails g alchemy:devise:install --force
-cd -
-BASH
+      Dir.chdir("spec/dummy") do
+        if ENV["ALCHEMY_BRANCH"] == "main"
+          system("bin/rails javascript:install:esbuild") || exit($?.exitstatus)
+        end
+        system(
+          <<~SETUP
+            bin/rake railties:install:migrations
+            bin/rake db:drop db:create db:migrate
+            bin/rails g alchemy:install --force --auto-accept
+            bin/rails g alchemy:devise:install --force
+          SETUP
+        )
+        exit($?.exitstatus) unless $?.success?
+        if ENV["ALCHEMY_BRANCH"] == "main"
+          system("bin/rails javascript:build") || exit($?.exitstatus)
+        else
+          system("bin/rails webpacker:compile") || exit($?.exitstatus)
+        end
+      end
     end
   end
 
   namespace :changelog do
     desc "Update changelog"
     task :update do
-      original_file = './CHANGELOG.md'
-      new_file = original_file + '.new'
-      backup = original_file + '.old'
-      changes = `git rev-list v#{ENV['PREVIOUS_VERSION']}...main | bundle exec github_fast_changelog AlchemyCMS/alchemy-devise`
-      File.open(new_file, 'w') do |fo|
+      original_file = "./CHANGELOG.md"
+      new_file = original_file + ".new"
+      backup = original_file + ".old"
+      changes = `git rev-list v#{ENV["PREVIOUS_VERSION"]}...main | bundle exec github_fast_changelog AlchemyCMS/alchemy-devise`
+      File.open(new_file, "w") do |fo|
         fo.puts changes
         File.foreach(original_file) do |li|
           fo.puts li
